refactor(agent): extract landing constants and media query hook
Extract `CONVERSATION_STARTERS`, `ALL_QUICK_SUGGESTIONS`, and `useMediaQuery` into dedicated modules to reduce file sizes and improve maintainability.
