+++ console/src/pages/agents/components/landing-constants.ts
@@ -1,0 +1,36 @@
+export const CONVERSATION_STARTERS = [
+  'How can I help secure your application today?',
+  'Ready to strengthen your security posture.',
+  'What security challenge are you facing?',
+  'Let me help you find and fix vulnerabilities.',
+  'Ask me anything about application security.',
+  'Your AI security assistant is ready.',
+  'Need help with a security issue?',
+  "What's your security concern today?",
+  "I'm listening. Tell me about your security needs.",
+  'Your security matters. I am here to help.',
+  'Share your security concerns. I am all ears.',
+  'I am ready to assist with your security questions.',
+  'Tell me what is on your mind regarding security.',
+  'I am here and attentive to your security needs.',
+  'Your security questions are important to me.',
+  'I am tuned in. What security topics interest you?',
+];
+
+export const ALL_QUICK_SUGGESTIONS = [
+  'What is the attack surface of my current workspace?',
+  'Are there any exposed services or ports in my project?',
+  'What security risks exist in my current environment?',
+  'How can I reduce the attack surface of my application?',
+  'What entry points need security hardening in my workspace?',
+  'Which network services are publicly accessible in my setup?',
+  'What APIs are exposed without proper authentication?',
+  'Are there unnecessary ports open in my infrastructure?',
+  'What cloud resources are vulnerable to external attacks?',
+  'How can I identify shadow IT in my organization?',
+  'What third-party integrations increase my attack surface?',
+  'Are there misconfigured security groups in my environment?',
+  'What endpoints lack proper rate limiting?',
+  'How can I discover undocumented API endpoints?',
+  'What services are running with excessive permissions?',
+];
